backend/requirements.txt
click==8.1.8
pytesseract==0.3.10
Pillow==10.2.0
pdf2image==1.17.0
python-dotenv==1.0.1
google-generativeai==0.3.2
tqdm==4.66.2
openai>=1.12.0
Flask>=2.0.0 